The ultimate merge

+++ flask/project/website/views.py
@@ -7,9 +7,12 @@ from project.controllers.desktop import DesktopController
 from project.controllers.tablet import TabletController
 from project.controllers.monitor import MonitorController
 from project.controllers.laptop import LaptopController
+from project.controllers.cart import CartController
+from project.controllers.purchase import PurchaseController
 from flask_login import login_required
 from project import logger
 from project.controllers.inventory import InventoryController
+import json
 
 website_blueprint = Blueprint('website_blueprint', __name__)
 
@@ -31,17 +34,18 @@ def index():
 
 @website_blueprint.route('/add-inventory/<string:electronic>/<string:model>', methods=['POST'])
 @login_required
-def add_desktop_inventory(electronic, model):
+def add_to_inventory(electronic, model):
     if g.user.admin:
-        InventoryController.add_item_to_inventory(model)
-        return redirect('/' + electronic)
+        InventoryController.add_item_to_inventory(model, electronic)
+        return redirect('/' + electronic.lower())
 
 
 @website_blueprint.route('/remove-inventory/<string:electronic>/<string:model>', methods=['POST'])
 @login_required
 def delete_item_from_inventory(electronic, model):
-    InventoryController.delete_item_from_inventory(model)
-    return redirect('/' + electronic)
+    if g.user.admin:
+        InventoryController.delete_item_from_inventory(model)
+        return redirect('/' + electronic.lower())
 
 
 @website_blueprint.route('/desktop', methods=['GET', 'POST'])
@@ -71,8 +75,36 @@ def desktop():
 @website_blueprint.route('/desktop-client', methods=['GET', 'POST'])
 @login_required
 def desktop_client():
-    return render_template('desktop-client.html', user=g.user, desktops=DesktopController.get_all_desktops())
+    desktops=DesktopController.get_all_desktops()
+    
+    filters = {
+        'brand': [],
+        'processor': [],
+        'ram_size': [],
+        'cpu_cores': []
+    }
 
+    for desktop in desktops:
+        if desktop[0]['brand'] not in filters['brand']: 
+            filters['brand'].append(desktop[0]['brand'])
+
+        if desktop[0]['processor'] not in filters['processor']: 
+            filters['processor'].append(desktop[0]['processor'])
+
+        if desktop[0]['ram_size'] not in filters['ram_size']: 
+            filters['ram_size'].append(desktop[0]['ram_size'])
+
+        if desktop[0]['cpu_cores'] not in filters['cpu_cores']: 
+            filters['cpu_cores'].append(desktop[0]['cpu_cores'])
+        
+    return render_template('desktop-client.html', user=g.user, filters=filters)
+
+@website_blueprint.route('/desktop-client/table', methods=['GET'])
+def desktop_client_table():    
+    logger.error(request.args.to_dict())
+    return json.dumps({
+        'data': DesktopController.get_all_unlocked_desktops(request.args.to_dict())
+    })
 
 @website_blueprint.route('/edit-desktop', methods=['POST'])
 @login_required
@@ -89,7 +121,7 @@ def edit_desktop():
         dimensions = request.form.get('desktopdimensions')
 
         if model and price and weight and brand and processor and ramsize and cpucores and hdsize and dimensions:
-            desktop = DesktopController.update_desktop(model, brand=brand, price=price, weight=weight, processor=processor, ram_size=ramsize, cpu_cores=cpucores, hd_size=hdsize, dimensions=dimensions)
+            desktop = DesktopController.update_desktop(model=model, brand=brand, price=price, weight=weight, processor=processor, ram_size=ramsize, cpu_cores=cpucores, hd_size=hdsize, dimensions=dimensions, hide=False)
             if desktop:
                 return redirect('/desktop')
             else:
@@ -97,6 +129,12 @@ def edit_desktop():
 
         # return render_template('desktop.html', user=g.user, desktops=DesktopController.get_all_desktops())
 
+@website_blueprint.route('/delete-desktop/<string:model>', methods=['POST'])
+@login_required
+def delete_desktop(model):
+    if g.user.admin:
+        DesktopController.delete_model(model)
+        return redirect('/desktop')
 
 @website_blueprint.route('/laptop', methods=['GET', 'POST'])
 @login_required
@@ -138,9 +176,35 @@ def laptop():
 @website_blueprint.route('/laptop-client', methods=['GET', 'POST'])
 @login_required
 def laptop_client():
-    return render_template('laptop-client.html', user=g.user, laptops=LaptopController.get_all_laptops())
+    laptops=LaptopController.get_all_laptops()
 
+    filters = {
+        'brand': [],
+        'processor': [],
+        'os': [],
+        'cpu_cores': []
+    }
 
+    for laptop in laptops:
+        if laptop[0]['brand'] not in filters['brand']: 
+            filters['brand'].append(laptop[0]['brand'])
+
+        if laptop[0]['processor'] not in filters['processor']: 
+            filters['processor'].append(laptop[0]['processor'])
+
+        if laptop[0]['os'] not in filters['os']: 
+            filters['os'].append(laptop[0]['os'])
+
+        if laptop[0]['cpu_cores'] not in filters['cpu_cores']: 
+            filters['cpu_cores'].append(laptop[0]['cpu_cores'])
+
+    return render_template('laptop-client.html', user=g.user, filters=filters)
+
+@website_blueprint.route('/laptop-client/table', methods=['GET'])
+def laptop_client_table():    
+    return json.dumps({
+        'data': LaptopController.get_all_unlocked_laptops(request.args.to_dict())
+    })
 
 @website_blueprint.route('/edit-laptop', methods=['POST'])
 @login_required
@@ -169,17 +233,22 @@ def edit_laptop():
             camera = False
 
         if model and price and weight and brand and processor and ramsize and cpucores and hdsize and displaysize:
-            laptop = LaptopController.update_laptop(model, brand=brand, price=price, weight=weight, display_size =displaysize,
+            laptop = LaptopController.update_laptop(model=model, brand=brand, price=price, weight=weight, display_size =displaysize,
                                                     processor=processor, ram_size=ramsize, cpu_cores=cpucores, hd_size=hdsize,
                                                     battery_info=battery, os=operatingsystem, touchscreen=touchscreen,
-                                                    camera=camera)
+                                                    camera=camera, hide=False)
 
             if laptop:
                 return redirect('/laptop')
             else:
                 logger.error('could not update laptop item')
 
-
+@website_blueprint.route('/delete-laptop/<string:model>', methods=['POST'])
+@login_required
+def delete_laptop(model):
+    if g.user.admin:
+        LaptopController.delete_model(model)
+        return redirect('/laptop')
 
 @website_blueprint.route('/tablet', methods=['GET', 'POST'])
 @login_required
@@ -213,8 +282,35 @@ def tablet():
 @website_blueprint.route('/tablet-client', methods=['GET', 'POST'])
 @login_required
 def tablet_client():
-    return render_template('tablet-client.html', user=g.user, tablets=TabletController.get_all_tablets())
+    tablets=TabletController.get_all_unlocked_tablets()
 
+    filters = {
+        'brand': [],
+        'os': [],
+        'dimensions': [],
+        'cpu_cores': []
+    }
+
+    for tablet in tablets:
+        if tablet[0]['brand'] not in filters['brand']: 
+            filters['brand'].append(tablet[0]['brand'])
+
+        if tablet[0]['os'] not in filters['os']: 
+            filters['os'].append(tablet[0]['os'])
+
+        if tablet[0]['dimensions'] not in filters['dimensions']: 
+            filters['dimensions'].append(tablet[0]['dimensions'])
+
+        if tablet[0]['cpu_cores'] not in filters['cpu_cores']: 
+            filters['cpu_cores'].append(tablet[0]['cpu_cores'])
+
+    return render_template('tablet-client.html', user=g.user, filters=filters)
+
+@website_blueprint.route('/tablet-client/table', methods=['GET'])
+def tablet_client_table():
+    return json.dumps({
+        'data': TabletController.get_all_unlocked_tablets(request.args.to_dict())
+    })
 
 @website_blueprint.route('/edit-tablet', methods=['POST'])
 @login_required
@@ -235,7 +331,7 @@ def edit_tablet():
         dimensions = request.form.get('dimensions')
 
         if model and price and weight and brand and processor and ramsize and cpucores and hdsize and displaysize:
-            tablet = TabletController.update_tablet(model, brand=brand , price=price, weight=weight, display_size=displaysize,
+            tablet = TabletController.update_tablet(model=model, brand=brand , price=price, weight=weight, display_size=displaysize,
                                                     dimensions=dimensions, processor=processor, ram_size=ramsize,
                                                     cpu_cores=cpucores, hd_size=hdsize, battery=battery, os=operatingsystem,
                                                     camera_info=camera)
@@ -244,6 +340,12 @@ def edit_tablet():
             else:
                 logger.error('couldnt create tablet item')
 
+@website_blueprint.route('/delete-tablet/<string:model>', methods=['POST'])
+@login_required
+def delete_tablet(model):
+    if g.user.admin:
+        TabletController.delete_model(model)
+        return redirect('/tablet')
 
 @website_blueprint.route('/monitor', methods=['GET', 'POST'])
 @login_required
@@ -270,18 +372,63 @@ def monitor():
 @website_blueprint.route('/monitor-client', methods=['GET', 'POST'])
 @login_required
 def monitor_client():
-    return render_template('monitor-client.html', user=g.user, monitors=MonitorController.get_all_monitors())
+    monitors=MonitorController.get_all_monitors()
+
+    filters = {
+        'brand': [],
+        'dimensions': []
+    }
+
+    for monitor in monitors:
+        if monitor[0]['brand'] not in filters['brand']: 
+            filters['brand'].append(monitor[0]['brand'])
+        if monitor[0]['dimensions'] not in filters['dimensions']: 
+            filters['dimensions'].append(monitor[0]['dimensions'])
+
+    return render_template('monitor-client.html', user=g.user, filters=filters)
+
+@website_blueprint.route('/monitor-client/table', methods=['GET'])
+def monitor_client_table():    
+    logger.error(request.args.to_dict())
+    return json.dumps({
+        'data': MonitorController.get_all_unlocked_monitors(request.args.to_dict())
+    })
 
 @website_blueprint.route('/cart', methods=['GET', 'POST'])
 @login_required
 def cart():
-    return render_template('cart.html', user=g.user)
+    return render_template('cart.html', user=g.user, items=CartController.get_cart_items())
+
+@website_blueprint.route('/add-to-cart/<string:model>/<string:item>', methods=['GET'])
+@login_required
+def add_to_cart(model, item):
+    CartController.add_item_to_cart(model)
+    return redirect('/' + item)
+
+
+@website_blueprint.route('/remove-from-cart/<string:model>', methods=['GET'])
+@login_required
+def remove_from_cart(model):
+    logger.critical('something')
+    CartController.remove_item_from_cart(model)
+    return redirect('/cart')
+
+@website_blueprint.route('/checkout/', methods=['GET'])
+@login_required
+def checkout_from_cart():
+    CartController.checkout_from_cart()
+    return redirect('/')
 
 @website_blueprint.route('/returns', methods=['GET', 'POST'])
 @login_required
 def returns():
-    return render_template('returns.html', user=g.user)
+    return render_template('returns.html', user=g.user, returns=PurchaseController.get_past_purchases())
 
+@website_blueprint.route('/return-item/<string:model>', methods=['GET', 'POST'])
+@login_required
+def return_past_purchase(model):
+    PurchaseController.return_item(model)
+    return redirect('/returns')
 
 @website_blueprint.route('/edit-monitor', methods=['POST'])
 @login_required
@@ -293,13 +440,18 @@ def edit_monitor():
         brand = request.form.get('brand')
         dimensions = request.form.get('monitor_dimensions')
         if model and price and weight and brand and dimensions:
-            monitor = MonitorController.update_monitor(model, brand=brand, price=price, weight=weight, dimensions=dimensions)
+            monitor = MonitorController.update_monitor(model=model, brand=brand, price=price, weight=weight, dimensions=dimensions)
             if monitor:
                 return redirect('/monitor')
             else:
                 logger.error('couldnt create monitor item')
 
-
+@website_blueprint.route('/delete-monitor/<string:model>', methods=['POST'])
+@login_required
+def delete_monitor(model):
+    if g.user.admin:
+        MonitorController.delete_model(model)
+        return redirect('/monitor')
 
 @website_blueprint.route('/dashboard', methods=['GET', 'POST'])
 @login_required
